Whoops

+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/profile/NostalgiaProfileCreator.kt
@@ -136,7 +136,7 @@ class NostalgiaProfileCreator : ProfileCreator {
 
 			val marrySection = ImageIO.read(File(Loritta.ASSETS, "profile/nostalgia/marry.png"))
 			graphics.drawImage(marrySection, 0, 0, null)
-			val marriedWith = runBlocking { lorittaShards.retrieveUserInfoById(marriedWithId) }
+			val marriedWith = runBlocking { lorittaShards.retrieveUserInfoById(marriedWithId.toLong()) }
 
 			if (marriedWith != null) {
 				val whitneySemiBold16 = whitneySemiBold.deriveFont(16f)
